requirements.txt
streamlit
openai
google-generativeai
python-dotenv
pillow
python-docx
PyPDF2